# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: NewsHub API
  description: Modern news platform backend API
  version: 1.0.0
  contact:
    name: NewsHub Team
    email: contact@newshub.com
servers:
  - url: http://localhost:3001
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Article:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        author:
          type: string
        lang:
          type: string
        sourceId:
          type: string
        status:
          type: string
          enum: [draft, published, rejected]
        aiInfo:
          type: object
          properties:
            rewritten:
              type: boolean
            plagiarismScore:
              type: number
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /articles:
    get:
      summary: Get list of articles
      tags: [Public]
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: lang
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, popular, trending]
            default: latest
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: List of articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer

    post:
      summary: Create a new article draft
      tags: [Admin/Editor]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
      responses:
        201:
          description: Article created successfully
        401:
          description: Unauthorized
  /articles/{slug}:
    get:
      summary: Get single article by slug
      tags: [Public]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Article details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        404:
          description: Article not found
...